more strong ASHE

+++ crypto/ASHE.hh
@@ -6,7 +6,7 @@ class ASHE{
     static const unsigned long long ASHE_MAX;
     std::string key;
     blowfish bf;
-    int IV;
+    uint64_t IV;
 public:
     ASHE(std::string s,int i);
     long encrypt(unsigned long long plaintext);
